# docs/openapi.yaml
openapi: 3.0.1
info:
  title: L0 Auth
  description: Authorization Service
  license:
    name: MIT
    url: https://github.com/tiki/l0-auth/blob/main/LICENSE
  version: 0.2.0
servers:
- url: https://auth.l0.mytiki.com
paths:
...
  /api/latest/app/{appId}:
    get:
      tags:
      - APP
      summary: Get App
      description: Get an app's profile
      operationId: l0-auth-app-get
      parameters:
      - name: appId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppInfoAO'
      security:
      - jwt: []
    post:
      tags:
      - APP
      summary: Update App
      description: Create a new App
      operationId: l0-auth-app-update
      parameters:
      - name: appId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppInfoAOReq'
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppInfoAO'
      security:
      - jwt: []
    delete:
      tags:
      - APP
      summary: Delete App
      description: Delete an App
      operationId: l0-auth-app-delete
      parameters:
      - name: appId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
      security:
      - jwt: []
...
components:
  schemas:
...
    AppInfoAOReq:
      type: object
      properties:
        name:
          type: string
    AppInfoAO:
      type: object
      properties:
        appId:
          type: string
        name:
          type: string
        users:
          uniqueItems: true
          type: array
          items:
            type: string
        modified:
          type: string
          format: date-time
        created:
          type: string
          format: date-time
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT